requirements.txt
fastapi
uvicorn
semantic-kernel
redis
pyodbc
azure-cosmos
slowapi
aiohttp
python-dotenv
